requirements.txt
langchain==0.3.20
langchain-community==0.3.20
langchain-huggingface==0.1.2
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
groq==0.13.2
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
python-dotenv==1.0.0